requirements.txt
absl-py==2.1.0
accelerate==0.29.1
# aiofiles==24.1.0
aiohttp==3.9.3
aiolimiter==1.1.0
aiosignal==1.3.1
alabaster==0.7.16
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
anytree==2.12.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autograd==1.7.0
autopep8==2.1.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-search-documents==11.5.2
azure-storage-blob==12.24.0
Babel==2.14.0
beartype==0.18.5
bitsandbytes==0.43.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
datasets==2.20.0
datashaper==0.0.49
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
devtools==0.12.2
# dgl==2.1.0+cu121
https://data.dgl.ai/wheels/cu121/dgl-2.1.0%2Bcu121-cp310-cp310-manylinux1_x86_64.whl
dglgo==0.0.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docutils==0.20.1
eagerpy==0.30.0
easydict==1.13 #
environs==11.2.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.5
ffmpy==0.3.2
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.51.0
foolbox==3.3.4
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.5.4
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
# gradio==5.6.0
# gradio==5.24.0
# gradio_client==1.4.3
# gradio_client==1.8.0
graphrag==0.5.0
graspologic==3.4.1
graspologic-native==1.2.1
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.2
hyppo==0.4.0
idna==3.6
igraph==0.11.4
imagesize==1.4.1
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.4
ipython==8.23.0
isodate==0.7.2
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.7.1
joblib==1.4.2
json_repair==0.30.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.1.1
kiwisolver==1.4.5
lancedb==0.13.0
libclang==18.1.1
littleutils==0.2.2
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
numpydoc==1.7.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
ogb==1.3.6
openai==1.54.5
openapi==2.0.0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.0
outdated==0.2.2
overrides==7.7.0
packaging==24.0
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
peft @ git+https://github.com/huggingface/peft.git@13e53fc7ee5d89d59b16523051006dddf0fb7a49
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
portalocker==2.10.1
POT==0.9.5
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyaml-env==1.2.1
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.10.0
pylance==0.17.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
rdkit-pypi==2022.9.5
referencing==0.34.0
regex==2023.12.25
requests==2.32.3
retry==0.9.2
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.5
safehttpx==0.1.1
safetensors==0.4.2
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.4.0
tiktoken==0.7.0
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
# torch==2.1.2+cu121
https://download.pytorch.org/whl/cu121/torch-2.1.2%2Bcu121-cp310-cp310-linux_x86_64.whl
torch_geometric==2.5.2
# torchaudio==2.1.2+cu121
https://download.pytorch.org/whl/cu121/torchaudio-2.1.2%2Bcu121-cp310-cp310-linux_x86_64.whl
torchdata==0.7.1
# torchvision==0.16.2+cu121
https://download.pytorch.org/whl/cu121/torchvision-0.16.2%2Bcu121-cp310-cp310-linux_x86_64.whl
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
transformers==4.39.3
triton==2.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.7
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
